made some small file structure changes and left join hashmaps challenge completed with tests
