+++ dataStructures/lib/src/main/java/codechallenge/leftJoinHashMaps/README.md
@@ -1,0 +1,21 @@
+# Left Join Hashmaps Challenge
+
+## Description
+
+In the world of databases, a "left join" retrieves full records from one table along with matching records (where present) from another table. If no match is found, the result is rendered as NULL. Adapting this concept for hashmaps, the `LeftJoinHashMaps` challenge seeks to merge two hashmaps in line with the LEFT JOIN principle.
+
+The main task can be found within the `LeftJoinHashMaps.java` file, located at: `./dataStructures/lib/src/main/java/codechallenge/leftJoinHashMaps/LeftJoinHashMaps.java`
+
+---
+
+## How to Run the Test
+
+For executing the tests related to the `leftJoinHashmaps` challenge:
+
+1. Ensure your Java project has been properly set up with JUnit.
+2. Download or clone this repository to your system.
+3. Open the project using your Java IDE of choice (options include Eclipse, IntelliJ IDEA, Visual Studio Code, and others).
+4. Navigate to the `LeftJoinHashMapsTest.java` file, located here: `./dataStructures/lib/src/test/java/codeChallenge/leftJoinHashMaps/LeftJoinHashMapsTest.java`.
+5. Upon right-clicking the `LeftJoinHashMapsTest.java` file, select either "Run As" or "Run Test" based on your IDE's terminology.
+
+---
