--- dataStructures/lib/src/main/java/datastructures/hashmap/HashMap.java
@@ -1,64 +1,0 @@
-package datastructures.hashmap;
-
-
-public class HashMap<K, V> {
-
-    private static final int DEFAULT_CAPACITY = 16;
-    private Node<K, V>[] table;
-    private int size;
-
-    public HashMap() {
-        table = new Node[DEFAULT_CAPACITY];
-    }
-
-    public void put(K key, V value) {
-        int index = getIndex(key);
-        Node<K, V> newNode = new Node<>(key, value);
-        if (table[index] == null) {
-            table[index] = newNode;
-        } else {
-            Node<K, V> current = table[index];
-            while (current.next != null && !current.key.equals(key)) {
-                current = current.next;
-            }
-            if (current.key.equals(key)) {
-                current.value = value;
-            } else {
-                current.next = newNode;
-            }
-        }
-        size++;
-    }
-
-    public V get(K key) {
-        int index = getIndex(key);
-        Node<K, V> current = table[index];
-        while (current != null) {
-            if (current.key.equals(key)) {
-                return current.value;
-            }
-            current = current.next;
-        }
-        return null;
-    }
-
-    private int getIndex(K key) {
-        return (key.hashCode() % table.length + table.length) % table.length;
-    }
-
-    public V getOrDefault(K key, V defaultValue) {
-        V value = get(key);
-        return value != null ? value : defaultValue;
-    }
-
-    private static class Node<K, V> {
-        K key;
-        V value;
-        Node<K, V> next;
-
-        Node(K key, V value) {
-            this.key = key;
-            this.value = value;
-        }
-    }
-}
