--- dataStructures/lib/src/main/java/datastructures/hashmap/HashMapChallenges.java
@@ -1,23 +1,0 @@
-package datastructures.hashmap;
-
-public class HashMapChallenges {
-    public String getFirstRepeatedWord(String str) {
-        if (str == null || str.isEmpty()) {
-            return "";
-        }
-
-        HashMap<String, Integer> wordMap = new HashMap<>();
-        String[] words = str.toLowerCase().split("\\W+");
-        // Split string into words using non-word characters but has issues with apostrophes
-
-        for (String word : words) {
-            int count = wordMap.getOrDefault(word, 0);
-            if (count == 1) {
-                return word;
-            }
-            wordMap.put(word, count + 1);
-        }
-
-        return "";
-    }
-}
